requirements.txt
absl-py
apache-beam
chex
flax
gensim
gin
git+https://github.com/google-research/t5x.git
git+https://github.com/google/flaxformer.git
grain
jax
jiwer
matplotlib
ml_collections
networkx
nltk
numpy
optax
pandas
protobuf
pycairo
pynini
pytest
scipy
seaborn
seqio
svgpathtools
t5
tensorflow
